backend/requirements.txt
fastapi
uvicorn
pydantic
gitpython
pytest
pylint
docker
requests
langgraph
langchain
langchain_google_genai
